+++ XuanYu.Editor.UI/Left/Left.axaml.cs
@@ -1,5 +1,4 @@
 using Avalonia.Controls;
-using Avalonia.Interactivity;
 using XYUI.Avalonia.Controls;
 
 namespace XuanYu.Editor.UI;
@@ -12,15 +11,12 @@ public partial class Left : UserControl
         SelectContentTab("project");
     }
 
-    void ProjectToggle_Click(object? sender, RoutedEventArgs e) => SelectContentTab("project");
-
-    void FileToggle_Click(object? sender, RoutedEventArgs e) => SelectContentTab("file");
+    void ContentTabs_SelectionChanged(object? sender, XYTab tab) => SelectContentTab(tab.Id);
 
     void SelectContentTab(string id)
     {
         var project = id != "file";
-        ProjectToggle.IsChecked = project;
-        FileToggle.IsChecked = !project;
+        ContentTabs.SelectedTabId = project ? "project" : "file";
         ProjectWorkspace.IsVisible = project;
         FileWorkspace.IsVisible = !project;
     }